requirements.txt
kafka-python
psycopg2-binary
requests
pandas
=streamlit
